Refactored tests: Moved datatype-related stream tests to LobsStreaming test, removed redundant cleanup logging, and handled table cleanup in @AfterEach.
